Code cleanup: PHPlot should not define or affect anything outside its own class.
- Removed the check for __FUNCTION__ (PHP 4.3 and up). This is obsolete.

- Do not set error_reporting to E_ALL. Although it is recommended that
scripts do this, it is not the place of loaded classes to do it.

- Remove unused global constant TOTY.

- Removed constants MAXY and MINY. Global constants like this are bad.
These were used as magic index values into data[] to hold min and max Y
values for the row. Instead, put them in separate arrays which are
named data_miny[] and data_maxy[]. (This seems to be only used by the
data line drawing function.)

Comment cleanup: Remove one commented-out partial function DrawPlotLabel,
and fix another commented-out code fragment in DrawYErrorBar. Both of these
had unmatched braces in them which caused a balance-braces check to fail.


git-svn-id: https://svn.code.sf.net/p/phplot/code/trunk@269 1589d9c6-5bb9-42a2-b2f8-3691a6793a8d