Allow mixing GD and TTF text on the same plot. Add SetFontGD() and SetFontTTF() methods, to set an element's font to a specific font type (GD or TTF) regardless of the SetUseTTF() setting.  SetUseTTF() now sets the default font type and still initializes all fonts to that type.
When SetFont() is used, rather than one of the new functions, the default
font type is assumed.  So the external interface is compatible, and
behavior should not change for existing plots.

Internals changes: all font data arrays are now kept in a single class
array variable $fonts.  For example, instead of $this->generic_font,
use $this->fonts['generic'].
Usage of internal functions ProcessTextGD() and ProcessTextTTF() changed.
These now take the font array as one argument, not split out.

Note: This change came from work done trying to fix TTF text positioning,
where it looks like additional information needs to be stored for TrueType
fonts. The old font data structure was awkward to extend, and allowing
mixed GD/TTF text was on the to-do list anyway.


Marked version in the source as being an unreleased CVS version.


git-svn-id: https://svn.code.sf.net/p/phplot/code/trunk@346 1589d9c6-5bb9-42a2-b2f8-3691a6793a8d